requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
opencv-python-headless
python-multipart
numpy
